backend POST for carbon done

+++ GreenRoutine-api/Models/UserChallenge.cs
@@ -14,5 +14,5 @@ public class UserChallenge
     public DateTime SignupDate { get; set; } = DateTime.Now;
 
     public string? Impact { get; set; }
-    public int Carbon_lb { get; set; }
+    public double Carbon_lb { get; set; }
 }
